Don't violate fn name constraints in fn init (#474)
The CLI allowed users to use function names that wouldn't be accepted by
the server. For example, names longer than 30 characters. Change the
code so that it uses the same validation code as the server to prevent
this from happening.

Furthermore, ensure the trigger name isn't violated when appending
'-trigger' to the function name.

Fixes #474